slight tidy up
